+++ src/GraphGenerator.h
@@ -12,28 +12,28 @@
 
 class GraphGenerator {
 public:
-    uint32_t n_A;
-    uint32_t n_B;
-    std::vector<uint32_t> allowed_n_connections;
+    int n_A;
+    int n_B;
+    std::vector<int> allowed_n_connections;
 
-    std::vector<uint32_t> edge_list;
+    std::vector<int> edge_list;
 
-    GraphGenerator(uint32_t n_A,
-                   uint32_t n_B,
-                   std::vector<uint32_t> allowed_n_connections) {
+    GraphGenerator(int n_A,
+                   int n_B,
+                   std::vector<int> allowed_n_connections) {
         this->n_A = n_A;
         this->n_B = n_B;
         this->allowed_n_connections = std::move(allowed_n_connections);
     }
 
     void generate() {
-        for (size_t i = 0; i < n_A; ++i) {
-            uint32_t vertex_a = i + 1;
-            uint32_t idx = random() % allowed_n_connections.size();
-            uint32_t n_edges = allowed_n_connections[idx];
+        for (int i = 0; i < n_A; ++i) {
+            int vertex_a = i + 1;
+            int idx = (int) random() % (int) allowed_n_connections.size();
+            int n_edges = allowed_n_connections[idx];
 
-            for (size_t j = 0; j < n_edges; ++j) {
-                uint32_t vertex_b = n_A + 1 + (random() % n_B);
+            for (int j = 0; j < n_edges; ++j) {
+                int vertex_b = n_A + 1 + ((int) random() % n_B);
 
                 bool edge_found = false;
                 for (size_t k = 0; k < edge_list.size() / 2; ++k) {
@@ -57,8 +57,8 @@ public:
         if (file.is_open()) {
             file << "p ocr " << n_A << " " << n_B << " " << edge_list.size() / 2 << "\n";
             for (size_t i = 0; i < edge_list.size() / 2; ++i) {
-                uint32_t vertex_a = edge_list[i * 2];
-                uint32_t vertex_b = edge_list[i * 2 + 1];
+                int vertex_a = edge_list[i * 2];
+                int vertex_b = edge_list[i * 2 + 1];
 
                 file << vertex_a << " " << vertex_b << "\n";
             }
